evoral/MIDIEvent: Whitespace cosmetics in header, make XML output work properly eg show the numbers instead of blanks
git-svn-id: svn://localhost/ardour2/branches/3.0@13221 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/evoral/src/MIDIEvent.cpp
@@ -60,15 +60,15 @@ MIDIEvent<Time>::to_xml() const
 	switch (type()) {
 	case MIDI_CMD_CONTROL:
 		result = new XMLNode("ControlChange");
-		result->add_property("Channel", channel());
-		result->add_property("Control", cc_number());
-		result->add_property("Value",   cc_value());
+		result->add_property("Channel", long(channel()));
+		result->add_property("Control", long(cc_number()));
+		result->add_property("Value",   long(cc_value()));
 		break;
 
 	case MIDI_CMD_PGM_CHANGE:
 		result = new XMLNode("ProgramChange");
-		result->add_property("Channel", channel());
-		result->add_property("Number",  pgm_number());
+		result->add_property("Channel", long(channel()));
+		result->add_property("Number",  long(pgm_number()));
 		break;
 
 	default:
